Moved the mock_moza fixture up to pytest's shared fixture location and split Xenbase tests up in their own package
